pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "color-analysis-tool"
version = "1.0.0"
description = "A comprehensive tool for analyzing colors in images with harmony calculations and multiple color space conversions"
readme = "README.md"
license = "MIT"
requires-python = ">=3.7"
authors = [
    { name = "Michail Semoglou" }
]
maintainers = [
    { name = "Michail Semoglou" }
]
keywords = [
    "color",
    "color analysis",
    "image processing",
    "color harmony",
    "design tools",
    "color extraction",
    "RGB",
    "CMYK",
    "HEX",
    "color space",
    "batch processing",
    "dominant color"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Visualization",
    "Topic :: Artistic Software",
    "Typing :: Typed"
]

dependencies = [
    "Pillow>=9.0.0",
    "tqdm>=4.65.0",
    "colormath>=3.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "build>=0.10.0",
    "twine>=4.0.0"
]

[project.urls]
Homepage = "https://github.com/MichailSemoglou/color-analysis-tool"
Repository = "https://github.com/MichailSemoglou/color-analysis-tool"
Documentation = "https://github.com/MichailSemoglou/color-analysis-tool#readme"
Issues = "https://github.com/MichailSemoglou/color-analysis-tool/issues"
Changelog = "https://github.com/MichailSemoglou/color-analysis-tool/blob/main/CHANGELOG.md"

[project.scripts]
color-analysis = "color_analysis_tool.cli:main"

[tool.hatch.build.targets.sdist]
include = [
    "/color_analysis_tool",
    "/LICENSE",
    "/README.md",
    "/CHANGELOG.md",
    "/CITATION.cff"
]

[tool.hatch.build.targets.wheel]
packages = ["color_analysis_tool"]

[tool.black]
line-length = 100
target-version = ['py37', 'py38', 'py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
